refactor: restructure ascent routes and contracts
