+++ src/server/contracts/contract.ts
@@ -1,0 +1,10 @@
+import { create, findById, list, search } from './ascents.ts'
+
+export const contract = {
+  ascents: {
+    create,
+    findById,
+    list,
+    search,
+  },
+}
