+++ src/server/routes/server.ts
@@ -1,0 +1,4 @@
+import { implement } from '@orpc/server'
+import { contract } from 'contracts/contract.ts'
+
+export const orpcServer = implement(contract)
